Big overhaul on saving: JSON files now stored in file system as opposed to database.

+++ scripts/data/index.php
@@ -2,9 +2,9 @@
 error_reporting(E_ALL);
 ini_set('display_errors', 1);
 require_once("dataconnect.php");
-require_once("deployment.php");
+require_once("fileconnect.php");
 $database = new database();
-$deployer = new deployer();
+$fileManager = new fileManager();
 
 if(isset($_GET['action']) and $_GET['action'] == "login")
 {
@@ -20,17 +20,28 @@ if(isset($_GET['action']) and $_GET['action'] == "getProjects")
   echo $database->getProjects($_GET);
   exit;
 }
-if(isset($_GET['action']) and $_GET['action'] == "saveProject")
+if(isset($_GET['action']) and $_GET['action'] == "newProject")
 {
   echo $database->saveProject($_GET);
   exit;
 }
+if(isset($_GET['action']) and $_GET['action'] == "saveProject")
+{
+  echo $fileManager->saveFile($_GET);
+  exit;
+}
+if(isset($_GET['action']) and $_GET['action'] == "loadProject")
+{
+  echo $fileManager->loadFile($_GET);
+  exit;
+}
 if(isset($_GET['action']) and $_GET['action'] == "deleteProject"){
   echo $database->deleteProject($_GET);
+  echo $fileManager->deleteFile($_GET);
   exit;
 }
 if(isset($_GET['action']) and $_GET['action'] == "deploy"){
-  echo $deployer->deploy($_GET);
+  echo $fileManager->deploy($_GET);
   exit;
 }
 ?>
